+++ client/views/nav/logo.js
@@ -1,0 +1,8 @@
+Template[getTemplate('logo')].helpers({
+  site_title: function(){
+    return getSetting('title', "Telescope");
+  },
+  logo_url: function(){
+    return getSetting('logoUrl');
+  }
+});