+++ client/views/nav/user_menu.js
@@ -1,14 +1,11 @@
 Template[getTemplate('userMenu')].helpers({
-  isLoggedIn: function () {
-    return !!Meteor.user();
-  },
-  name: function () {
-    return getDisplayName(Meteor.user());
-  },
-  profileUrl: function () {
-    return Router.path('user_profile', {_idOrSlug: Meteor.user().slug});
-  },
-  userEditUrl: function () {
-    return Router.path('user_edit', {slug: Meteor.user().slug});
+  userMenuData: function () {
+    return {
+      dropdownName: 'user',
+      dropdownLabel: getDisplayName(Meteor.user()),
+      dropdownItems: userMenu,
+      dropdownClass: 'header-submodule',
+      dropdownMode: getSetting('navLayout', 'top-nav') == 'top-nav' ? 'hover' : 'accordion'
+    }
   }
 });