+++ packages/telescope-base/lib/base.js
@@ -34,16 +34,8 @@ STATUS_REJECTED=3;
 // ------------------------------------- Navigation -------------------------------- //
 
 
-// array containing nav items; initialize with views menu and admin menu
+// array containing nav items;
 primaryNav = [
-  {
-    template: 'viewsMenu',
-    order: 10
-  },
-  {
-    template: 'adminMenu',
-    order: 20
-  }
 ];
 
 secondaryNav = [
@@ -58,30 +50,6 @@ secondaryNav = [
   {
     template: 'submitButton',
     order: 30
-  }
-];
-
-// array containing items in the admin menu
-adminMenu = [
-  {
-    route: 'posts_pending',
-    label: 'Pending',
-    description: 'posts_awaiting_moderation'
-  },
-  {
-    route: 'posts_scheduled',
-    label: 'Scheduled',
-    description: 'future_scheduled_posts'
-  },
-  {
-    route: 'all-users',
-    label: 'Users',
-    description: 'users_dashboard'
-  },
-  {
-    route: 'settings',
-    label: 'Settings',
-    description: 'telescope_settings_panel'
   }
 ];
 
@@ -101,9 +69,62 @@ viewsMenu = [
     route: 'posts_best',
     label: 'best',
     description: 'highest_ranked_posts_ever'
+  },
+  {
+    route: 'posts_pending',
+    label: 'pending',
+    description: 'posts_awaiting_moderation',
+    adminOnly: true
+  },
+  {
+    route: 'posts_scheduled',
+    label: 'scheduled',
+    description: 'future_scheduled_posts',
+    adminOnly: true
+  },  
+];
+
+// array containing items in the admin menu
+adminMenu = [
+  {
+    route: 'settings',
+    label: 'Settings',
+    description: 'telescope_settings_panel'
+  },
+  {
+    route: 'usersDashboard',
+    label: 'Users',
+    description: 'users_dashboard'
   }
 ];
 
+userMenu = [
+  {
+    route: function () {
+      return Router.path('user_profile', {_idOrSlug: Meteor.user().slug});
+    },
+    label: 'profile',
+    description: 'view_your_profile'
+  },
+  {
+    route: function () {
+      return Router.path('user_edit', {slug: Meteor.user().slug});
+    },
+    label: 'edit_account',
+    description: 'edit_your_profile'
+  },
+  {
+    route: 'settings',
+    label: 'settings',
+    description: 'settings',
+    adminOnly: true
+  },
+  {
+    route: 'signOut',
+    label: 'sign_out',
+    description: 'sign_out'
+  }
+]
 // ------------------------------------- Views -------------------------------- //
 
 
@@ -186,6 +207,13 @@ heroModules = [];
 footerModules = [];
 
 threadModules = [];
+
+postListTopModules = [
+  {
+    template: 'postViewsNav',
+    order: 99
+  }
+];
 
 postModules = [
   {
@@ -324,7 +352,6 @@ userProfileCompleteChecks.push(
 );
 
 // ------------------------------ Dynamic Templates ------------------------------ //
-
 
 templates = {}