+++ packages/telescope-singleday/lib/routes.js
@@ -4,6 +4,11 @@ PostsSingledayController = RouteController.extend({
 
   template: getTemplate('singleDay'),
 
+  onBeforeAction: function () {
+    this.render(getTemplate('postListTop'), {to: 'postListTop'});
+    this.next();
+  },
+  
   data: function() {
     var currentDate = this.params.day ? new Date(this.params.year, this.params.month-1, this.params.day) : Session.get('today');
     Session.set('currentDate', currentDate);