+++ packages/telescope-tagline-banner/lib/client/templates/tagline_banner.js
@@ -1,6 +1,6 @@
 Template[getTemplate('taglineBanner')].helpers({
   showTaglineBanner: function () {
-    return Router.current().location.get().path == '/' && !!getSetting('tagline') && !!getSetting('showTaglineBanner');
+    return !!getSetting('tagline') && !!getSetting('showTaglineBanner');
   }
 });