+++ packages/telescope-tagline-banner/lib/tagline.js
@@ -1,6 +1,6 @@
-heroModules.push({
+postListTopModules.push({
   template: 'taglineBanner',
-  order: 0
+  order: 1
 });
 
 var showTaglineBanner = {