+++ packages/telescope-tags/lib/client/templates/categories_menu.js
@@ -1,16 +1,19 @@
 Meteor.startup(function () {
   Template[getTemplate('categoriesMenu')].helpers({
-    hasCategories: function(){
-      return typeof Categories !== 'undefined' && Categories.find().count();
-    },
-    cat: function () {
-      return __('categories')
-    },
-    categories: function(){
-      return Categories.find({}, {sort: {order: 1, name: 1}});
-    },
-    categoryLink: function () {
-      return getCategoryUrl(this.slug);
+    categoriesMenuData: function () {
+      return {
+        dropdownName: 'categories',
+        dropdownItems: _.map(Categories.find({}, {sort: {order: 1, name: 1}}).fetch(), function (category) {
+          return {
+            route: function () {
+              return getCategoryUrl(category.slug);
+            },
+            label: category.name
+          }
+        }),
+        dropdownClass: 'header-submodule',
+        dropdownMode: getSetting('navLayout', 'top-nav') == 'top-nav' ? 'hover' : 'accordion'
+      }
     }
   });
 });